fix component configration
