+++ package.json
@@ -14,8 +14,7 @@
   "dependencies": {
     "@nuxtjs/axios": "^5.13.1",
     "@nuxtjs/dotenv": "^1.4.1",
-    "chart.js": "2.8",
-    "chartjs-plugin-annotation": "^0.5.7",
+    "chart.js": "^2.8.0",
     "core-js": "^3.9.1",
     "google-palette": "^1.1.0",
     "nuxt": "^2.15.3",
