+++ pages/index.vue
@@ -6,18 +6,18 @@
         v-for="prefecture in prefectures"
         :key="prefecture.prefCode"
         :prefecture="prefecture"
-        @check="getPopulation($event)"
+        @check="addPopulation($event)"
         @uncheck="deletePopulation($event)"
       />
     </div>
     <div class="small">
-      <PopulationChart :chart-data="datacollection" :options="options" />
+      <PopulationChart :checked-prefs="checkedPrefs" />
+      <p>※{{ border }}以降は推定値</p>
     </div>
   </div>
 </template>
 
 <script>
-import * as palette from 'google-palette'
 import PrefectureCheckbox from '~/components/PrefectureCheckbox.vue'
 import PopulationChart from '~/components/PopulationChart.vue'
 
@@ -28,58 +28,16 @@ export default {
   },
   async asyncData(context) {
     const result = await context.$axios.get('api/v1/prefectures')
-    console.log(result)
     return { prefectures: result.data.result }
   },
   data() {
     return {
       border: null,
       checkedPrefs: [],
-      datacollection: {
-        labels: [],
-        datasets: [],
-      },
-      options: {
-        responsive: true,
-        maintainAspectRatio: false,
-        scales: {
-          xAxes: [
-            {
-              scaleLabel: {
-                display: true,
-                labelString: '年度',
-              },
-            },
-          ],
-          yAxes: [
-            {
-              scaleLabel: {
-                display: true,
-                labelString: '人口数',
-              },
-            },
-          ],
-        },
-        annotation: {
-          annotations: [
-            {
-              type: 'box',
-              drawTime: 'afterDatasetsDraw',
-              xMin: 2, // 基準とするx軸の値（始点）
-              xMax: 3, // 基準とするx軸の値（終点）
-              yMin: 0, // 基準とするy軸の値（始点）
-              yMax: 0, // 基準とするy軸の値（終点）
-              backgroundColor: 'rgba(100,100,180,.2)',
-              borderWidth: 0,
-              borderColor: 'rgba(180,100,180,0)',
-            },
-          ],
-        },
-      },
     }
   },
   methods: {
-    async getPopulation(prefecture) {
+    async addPopulation(prefecture) {
       // apiから人口の情報を取得
       const result = await this.$axios.get(
         'api/v1/population/composition/perYear',
@@ -90,14 +48,16 @@ export default {
           },
         }
       )
-      this.border = result.boundaryYear
+      // 実績値と推計値の区切り年をセット
+      if (this.border === null) {
+        this.border = result.data.result.boundaryYear
+      }
       // checkedPrefsに県名と人口の情報を持ったオブジェクトを追加
       const prefPopulation = {
         prefName: prefecture.prefName,
         populationInfo: result.data.result,
       }
       this.checkedPrefs.push(prefPopulation)
-      this.updateDatacollection(this.checkedPrefs)
     },
     deletePopulation(targetPref) {
       // checkedPrefsからチェックを外した県を削除
@@ -106,31 +66,6 @@ export default {
           this.checkedPrefs.splice(index, 1)
         }
       })
-      this.updateDatacollection(this.checkedPrefs)
-    },
-    updateDatacollection(checkedPrefs) {
-      const newDatacollection = {}
-      newDatacollection.datasets = []
-      // グラフに割り当てる色を自動生成
-      const borderColors = palette('mpn65', this.checkedPrefs.length).map(
-        (hex) => {
-          return '#' + hex
-        }
-      )
-      checkedPrefs.forEach((checkedPref, index) => {
-        if (index === 0) {
-          newDatacollection.labels =
-            checkedPref.populationInfo.data[0].data.map((d) => d.year)
-        }
-        const chartData = {
-          label: checkedPref.prefName,
-          backgroundColor: 'rgba(0, 0, 0, 0)',
-          borderColor: borderColors[index],
-          data: checkedPref.populationInfo.data[0].data.map((d) => d.value),
-        }
-        newDatacollection.datasets.push(chartData)
-      })
-      this.datacollection = newDatacollection
     },
   },
 }
